chore: create ryot_legacy_assets crate
Moved all appearance-related functionalities from the main Ryot crate to the newly created ryot_legacy_assets crate.

This encapsulates all Tibia-specific asset handling, including sprite sheet decoding and appearances encoding/decoding integration, into a dedicated module.

This change streamlines the main Ryot crate and isolates game-specific asset management.
