+++ crates/ryot_internal/Cargo.toml
@@ -28,5 +28,6 @@ tiled = ["ryot_pathfinder?/tiled"]
 [dependencies]
 ryot_core = { path = "../ryot_core", version = "0.2.0" }
 ryot_grid = { path = "../ryot_grid", version = "0.2.0" }
+ryot_legacy_assets = { path = "../ryot_legacy_assets", version = "0.2.0", optional = true }
 ryot_pathfinder = { path = "../ryot_pathfinder", version = "0.2.0", optional = true }
