+++ crates/ryot_internal/src/prelude.rs
@@ -3,3 +3,6 @@ pub use ryot_grid::prelude::*;
 
 #[cfg(feature = "ryot_pathfinder")]
 pub use ryot_pathfinder::prelude::*;
+
+#[cfg(feature = "ryot_legacy_assets")]
+pub use ryot_legacy_assets::prelude::*;
