+++ crates/ryot_legacy_assets/build.rs
@@ -1,4 +1,5 @@
 use std::io::Result;
+
 fn main() -> Result<()> {
     prost_build::compile_protos(
         &["src/appearances/appearances.proto"],
