+++ crates/ryot_legacy_assets/src/appearances/tests.rs
